# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: LyoDex API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /operators:
    get:
      operationId: listOperators
      tags: [operators]
      summary: List all operators
      responses:
        "200":
          description: List of operators
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Operator"
    post:
      operationId: createOperator
      tags: [operators]
      summary: Create a new operator profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OperatorInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operator"
  /operators/me:
    get:
      operationId: getMyOperatorProfile
      tags: [operators]
      summary: Get the authenticated operator's own profile
      responses:
        "200":
          description: Operator profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operator"
        "404":
          description: No operator profile linked to this account
    post:
      operationId: createMyOperatorProfile
      tags: [operators]
      summary: Create an operator profile for the authenticated user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMyOperatorProfileInput"
      responses:
        "201":
          description: Created operator profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operator"
        "400":
          description: Validation error
        "409":
          description: Operator profile already exists for this account
    put:
      operationId: updateMyOperatorProfile
      tags: [operators]
      summary: Update the authenticated operator's own profile (non-admin fields only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OperatorProfileUpdate"
      responses:
        "200":
          description: Updated operator profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operator"
        "400":
          description: Validation error
        "404":
          description: No operator profile linked to this account
  /operators/{id}:
    get:
      operationId: getOperator
      tags: [operators]
      summary: Get an operator by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Operator found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operator"
        "404":
          description: Not found
...
components:
  schemas:
...
    Operator:
      type: object
      required: [id, name, location, capacity_kg, price_per_kg, certifications, verified_certifications, turnaround_days, rating, review_count, available, created_at]
      properties:
        id:
          type: integer
        name:
          type: string
        location:
          type: string
        description:
          type: ["string", "null"]
        capacity_kg:
          type: integer
        price_per_kg:
          type: number
        certifications:
          type: array
          items:
            type: string
        verified_certifications:
          type: array
          items:
            type: string
        cert_verified_at:
          type: object
          additionalProperties:
            type: string
        turnaround_days:
          type: integer
        rating:
          type: number
        review_count:
          type: integer
        available:
          type: boolean
        created_at:
          type: string
        verification_status:
          type: ["string", "null"]
          description: "'verified' | 'partially_verified' | 'not_verified'"
        source_urls:
          type: array
          items:
            type: string
        website_url:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        role:
          type: ["string", "null"]
          description: "'operator' | 'ingredient_seller' | 'machinery_supplier' | 'service_provider'"
        provider_types:
          type: array
          items:
            type: string
          description: "Multi-select service types: 'operator' | 'ingredient_seller' | 'service_provider'"
        last_verified_date:
          type: ["string", "null"]
        food_market_focus:
          type: ["boolean", "null"]
        pharmaceutical_focus:
          type: ["boolean", "null"]
        contact_page_url:
          type: ["string", "null"]
    OperatorInput:
      type: object
      required: [name, location, capacity_kg, price_per_kg, turnaround_days]
      properties:
        name:
          type: string
        location:
          type: string
        description:
          type: string
        capacity_kg:
          type: integer
        price_per_kg:
          type: number
        certifications:
          type: array
          items:
            type: string
        turnaround_days:
          type: integer
    CreateMyOperatorProfileInput:
      type: object
      description: Fields for creating a new operator profile. Name is required; numeric capacity/pricing fields default to 0/14 if omitted.
      required: [name]
      properties:
        name:
          type: string
        location:
          type: string
        description:
          type: ["string", "null"]
        capacity_kg:
          type: integer
        price_per_kg:
          type: number
        certifications:
          type: array
          items:
            type: string
        turnaround_days:
          type: integer
        available:
          type: boolean
        website_url:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        food_market_focus:
          type: ["boolean", "null"]
        pharmaceutical_focus:
          type: ["boolean", "null"]
        contact_page_url:
          type: ["string", "null"]
        gps_lat:
          type: ["number", "null"]
        gps_lng:
          type: ["number", "null"]
        service_radius_km:
          type: ["integer", "null"]
        contact_name:
          type: ["string", "null"]
        contact_email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
    OperatorProfileUpdate:
      type: object
      description: Fields an operator can self-edit. Sensitive admin-only fields (platform_fee_override, audit_status, verification_status) are excluded.
      properties:
        name:
          type: string
        location:
          type: string
        description:
          type: ["string", "null"]
        capacity_kg:
          type: integer
        price_per_kg:
          type: number
        certifications:
          type: array
          items:
            type: string
        turnaround_days:
          type: integer
        available:
          type: boolean
        website_url:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        food_market_focus:
          type: ["boolean", "null"]
        pharmaceutical_focus:
          type: ["boolean", "null"]
        contact_page_url:
          type: ["string", "null"]
        gps_lat:
          type: ["number", "null"]
        gps_lng:
          type: ["number", "null"]
        service_radius_km:
          type: ["integer", "null"]
        map_status:
          type: ["string", "null"]
          description: "'active' | 'inactive'"
        provider_types:
          type: array
          items:
            type: string
          description: "Selectable service types: 'operator', 'ingredient_seller', 'service_provider'"
        contact_name:
          type: ["string", "null"]
        contact_email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
...